# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Rooms Booking Service",
        "contact": {}
    },
    "host": "127.0.0.1:8080",
    "paths": {
        "/bookings/create": {
            "post": {
                "description": "Доступно только роли user. Администратор не может создавать брони (403). userId берётся из JWT-токена, а не из тела запроса. Если временной слот находится в прошлом (start \u003c now), возвращается 400 (INVALID_REQUEST).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Создать бронь на слот (только user). Опционально — запросить ссылку на конференцию.",
                "parameters": [
                    {
                        "description": "тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_bookings_transport_http.SlotDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Бронь создана",
                        "schema": {
                            "$ref": "#/definitions/internal_features_bookings_transport_http.BookingDTOResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён (бронирование доступно только роли user)",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "404": {
                        "description": "Слот не найден",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "409": {
                        "description": "Слот уже занят",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    }
                }
            }
        },
        "/bookings/list": {
            "get": {
                "description": "Доступно только роли admin. Поддерживает пагинацию через параметры page и pageSize. Оба параметра опциональны; значения по умолчанию: page=1, pageSize=20. Максимальное значение pageSize — 100.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Список всех броней с пагинацией (только admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Номер страницы (начиная с 1). По умолчанию 1.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Количество записей на странице. По умолчанию 20, максимум 100.",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cписок всех броней",
                        "schema": {
                            "$ref": "#/definitions/internal_features_bookings_transport_http.BookingsDTOWithPagination"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос (некорректные параметры пагинации)",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён (только admin)",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    }
                }
            }
        },
        "/bookings/my": {
            "get": {
                "description": "Доступно только роли user. Возвращает брони пользователя, чей user_id содержится в JWT-токене. Возвращаются только брони на будущие слоты (start \u003e= now); брони на уже прошедшие слоты в ответ не включаются.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Список броней текущего пользователя (только user)",
                "responses": {
                    "200": {
                        "description": "Список броней текущего пользователя",
                        "schema": {
                            "$ref": "#/definitions/internal_features_bookings_transport_http.BookingsDTOResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён (только user)",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.InternalError"
                        }
                    }
                }
            }
        },
        "/bookings/{bookingId}/cancel": {
            "post": {
                "description": "Доступно только роли user. Пользователь может отменить только свою бронь. Операция идемпотентна: повторный вызов на уже отменённой брони не является ошибкой и возвращает 200 с актуальным состоянием брони (status: cancelled). При попытке отменить чужую бронь — 403. При отсутствии брони с указанным ID — 404.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Отменить бронь (только своя бронь, только user)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор брони",
                        "name": "bookingId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Бронь отменена (или уже была отменена ранее)",
                        "schema": {
                            "$ref": "#/definitions/internal_features_bookings_transport_http.BookingDTOResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "403": {
                        "description": "Не своя бронь или роль не user",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "404": {
                        "description": "Бронь не найдена",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    }
                }
            }
        },
...
        "/rooms/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rooms"
                ],
                "summary": "Создать переговорку (только admin)",
                "parameters": [
                    {
                        "description": "тело запроса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_features_rooms_transport_http.CreateRoomDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Переговорка создана",
                        "schema": {
                            "$ref": "#/definitions/internal_features_rooms_transport_http.RoomDTOResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещён (требуется роль admin)",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.CustomError"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/backend-assignment-avito_internal_core_domain.InternalError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "backend-assignment-avito_internal_core_domain.CustomError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "enum": [
                        "INVALID_REQUEST",
                        " INTERNAL_ERROR",
                        " SCHEDULE_EXISTS",
                        " SLOT_ALREADY_BOOKED",
                        " FORBIDEN",
                        " NOT_FOUND"
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "invalid request"
                }
            }
        },
        "backend-assignment-avito_internal_core_domain.InternalError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "enum": [
                        "INTERNAL_ERROR"
                    ]
                },
                "message": {
                    "type": "string",
                    "example": "internal server error"
                }
            }
        },
...
        "internal_features_bookings_transport_http.BookingDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "0",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "slotId": {
                    "type": "string",
                    "x-order": "1",
                    "example": "4hbdaf64-5717-4562-b3fc-2c963f66aba6"
                },
                "userId": {
                    "type": "string",
                    "x-order": "3",
                    "example": "9hbd88864-5717-4562-b3fc-2c963f66af77"
                },
                "status": {
                    "type": "string",
                    "x-order": "4",
                    "example": "string"
                },
                "createdAt": {
                    "type": "string",
                    "x-order": "5",
                    "example": "2026-03-25T12:00:41.267Z"
                }
            }
        },
        "internal_features_bookings_transport_http.BookingDTOResponse": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/internal_features_bookings_transport_http.BookingDTO"
                }
            }
        },
        "internal_features_bookings_transport_http.BookingsDTOResponse": {
            "type": "object",
            "properties": {
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_features_bookings_transport_http.BookingDTO"
                    }
                }
            }
        },
        "internal_features_bookings_transport_http.BookingsDTOWithPagination": {
            "type": "object",
            "properties": {
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_features_bookings_transport_http.BookingDTO"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/internal_features_bookings_transport_http.PaginationDTO"
                }
            }
        },
        "internal_features_bookings_transport_http.PaginationDTO": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer",
                    "x-order": "0",
                    "example": 1
                },
                "pageSize": {
                    "type": "integer",
                    "x-order": "1",
                    "example": 10
                },
                "total": {
                    "type": "integer",
                    "x-order": "2",
                    "example": 15
                }
            }
        },
        "internal_features_bookings_transport_http.SlotDTO": {
            "type": "object",
            "properties": {
                "slotId": {
                    "type": "string"
                }
            }
        },
        "internal_features_rooms_transport_http.CreateRoomDTO": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "x-order": "0",
                    "example": "work_room"
                },
                "description": {
                    "type": "string",
                    "x-order": "1",
                    "example": "some description..."
                },
                "capacity": {
                    "type": "integer",
                    "x-order": "2",
                    "example": 3
                }
            }
        },
        "internal_features_rooms_transport_http.RoomDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "0",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "name": {
                    "type": "string",
                    "x-order": "1",
                    "example": "work_room"
                },
                "description": {
                    "type": "string",
                    "x-order": "2",
                    "example": "room for work..."
                },
                "capacity": {
                    "type": "integer",
                    "x-order": "3",
                    "example": 3
                },
                "createdAt": {
                    "type": "string",
                    "x-order": "4",
                    "example": "2026-03-25T12:00:41.267Z"
                }
            }
        },
        "internal_features_rooms_transport_http.RoomDTOResponse": {
            "type": "object",
            "properties": {
                "room": {
                    "$ref": "#/definitions/internal_features_rooms_transport_http.RoomDTO"
                }
            }
        },
...
    }
}